Stop paying for FP64 in a kernel that was never bandwidth-bound
The CUDA FP8 kernel accumulated in double, and dsv4_cuda.h justified it with
"these are bandwidth-bound kernels, so the arithmetic precision is close to
free". That was never measured, and it is false. Consumer Blackwell runs FP64
at a fraction of FP32, so the kernel was FP64-throughput-bound: on identical
memory traffic, dropping the inner loop to float is 3.3x.

Shipped is the middle option, which had not been tried: the dot product over
one 128-column scale block runs in float, and the scale multiply, the fold
across blocks and the cross-warp sum stay in double. That costs 2.5% against
pure float and keeps the error flat in the reduce dimension -- pure float's
grows, because the cross-block sum gets longer, and Pro reduces over 16384.

Interleaved A/B on the real Flash checkpoint, three pairs alternated on a
warming machine: attention 445 -> 309 ms/pass, wall clock 1.40 -> 1.23
s/token, and byte-identical token ids in all six runs. 21/21 gates green.

Two things fixed alongside, because the same measurement turned them up:

- The CUDA gate set every scale byte to 127, which is exactly 1.0, so the
  whole per-block scale multiply cancelled and the gate could not see a kernel
  that got the scale path wrong. Scales now vary over 2^-7..2^7.

- bench/gpu_call.c, and the README claim it kills. The page said there is "a
  ~1.3 ms floor per call that barely varies with size", inferred from two
  matrices that are BOTH 33.5 MB -- two points at one size cannot tell a fixed
  overhead from a cost linear in bytes. Fitting ten shapes over four orders of
  magnitude gives ms = 0.031 + 0.0150 * MB: the floor is 31 us, not 1.3 ms,
  and the call is bandwidth-bound nearly all the way down. The old number was
  the double kernel's run time being read as overhead.

The bench also reports, without explaining it, that a 512x512 CPU matmul takes
5.3 ms with 20 OpenMP threads and 0.41 ms with one, inside a process holding a
live CUDA context. Labelled as unexplained rather than left in the table
looking like a matmul time.

+++ src/cuda/dsv4_cuda.cu
@@ -68,6 +68,27 @@ __device__ __forceinline__ float d_e8m0(unsigned char b)
  * element -- matching inference/kernel.py's fp4_gemm and the CPU path. Scaling
  * each weight would be algebraically equal, numerically different, and cost a
  * multiply per element instead of per block.
+ *
+ * ACCUMULATION IS SPLIT, AND THE SPLIT IS MEASURED, NOT ASSUMED.
+ * The inner product over one 128-column scale block runs in float; the scale
+ * multiply, the fold across blocks and the final cross-warp sum run in double.
+ *
+ * This kernel used to be double throughout, justified by "bandwidth-bound, so
+ * precision is close to free". That was never measured and it is false --
+ * consumer Blackwell runs FP64 at a small fraction of FP32, and the shipped
+ * kernel was FP64-throughput-bound, not DRAM-bound. On identical memory
+ * traffic (RTX 5060 Laptop, 50 iterations, warm):
+ *
+ *     shape          double      float       hybrid    worst rel vs double
+ *     4096x8192      1.271 ms    0.373 ms    0.382 ms  float 4.3e-7  hyb 2.8e-7
+ *     4096x16384     2.538 ms    1.015 ms    1.040 ms  float 6.4e-7  hyb 2.8e-7
+ *     129280x4096   19.99  ms    7.365 ms    7.523 ms  float 8.2e-7  hyb 4.3e-7
+ *
+ * Hybrid costs 2-2.5% over pure float and buys back the property that matters:
+ * float's error grows with the reduce dimension because the cross-block sum
+ * gets longer, while the hybrid's does not -- only the fixed 128-element inner
+ * sum is narrow. Pro's reduce dimension is 16384, so that is not academic.
+ * Argmax was identical to the double kernel in every row of that table.
  */
 __global__ void k_mmq_fp8(float *__restrict__ y,
                           const float *__restrict__ x,
@@ -85,14 +106,14 @@ __global__ void k_mmq_fp8(float *__restrict__ y,
     double mine = 0.0;
     for (int c0 = warp * blk_c; c0 < in; c0 += nwarp * blk_c) {
         const int n = min(blk_c, in - c0);
-        double acc = 0.0;
+        float acc = 0.0f;
         for (int k = lane; k < n; k += 32)
-            acc = fma((double)d_e4m3(row[c0 + k]), (double)x[c0 + k], acc);
+            acc = fmaf(d_e4m3(row[c0 + k]), x[c0 + k], acc);
         /* Reduce the warp's 32 partials. */
         for (int off = 16; off > 0; off >>= 1)
             acc += __shfl_down_sync(0xFFFFFFFFu, acc, off);
         if (lane == 0)
-            mine += acc * (double)d_e8m0(S[(size_t)(o / blk_r) * sc
+            mine += (double)acc * (double)d_e8m0(S[(size_t)(o / blk_r) * sc
                                            + (c0 / blk_c)]);
     }
